+++ src/main.rs
@@ -62,6 +62,7 @@ use clap::{App, Arg};
 use std::env;
 
 pub mod coinbase;
+pub mod compiler;
 pub mod hashing;
 pub mod loader;
 pub mod middleware_result;
